+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.3] - 2024-07-04
+
+### Changed
+- `Timecard.calculateMinutesWorked()` now takes an optional second argument to include time since last entry to NOW.
+
+### Fixed
+- `Timecard.calculateMinutesWorked()` returning 0 minutes despite there being time worked
+
+
+
 ## [2.0.2] - 2024-07-03
 
 ### Changed